feat: 쿠폰 발급 요청 및 대기열 사용자 쿠폰 발급 처리 구현 (#146)
* style : Schedule 어노테이션 위치 변경

* refactor: 쿠폰 발행 기간 하루로 통일 및 쿠폰 정보 오픈 날짜 추가

* feat: 쿠폰 발행 가능 날짜 중복 체크 기능 추가

* refactor: Builder 삭제

* test: 쿠폰 관련 테스트 수정

* feat: 쿠폰 발행 관련 레포지토리 기능 구현 및 테스트

* test: 쿠폰 발행 관련 문자열 레디스 기능 구현 및 테스트

* feat: 쿠폰 발행 관련 ZSET 레디스 기능 구현 및 테스트

* test: 쿠폰 발행 컨트롤러 기능 테스트

* test: RestDoc 업데이트

* test: Github Actions 시, Redis ZSET 명령어 못찾는 테스트 Disable

+++ src/test/java/com/moabam/api/infrastructure/redis/StringRedisRepositoryTest.java
@@ -17,10 +17,11 @@ import com.moabam.global.config.EmbeddedRedisConfig;
 class StringRedisRepositoryTest {
 
 	@Autowired
-	private StringRedisRepository stringRedisRepository;
+	StringRedisRepository stringRedisRepository;
 
 	String key = "key";
 	String value = "value";
+	String stockKey = "key_INCR";
 	Duration duration = Duration.ofMillis(5000);
 
 	@BeforeEach
@@ -30,29 +31,25 @@ class StringRedisRepositoryTest {
 
 	@AfterEach
 	void setDown() {
-		stringRedisRepository.delete(key);
+		if (stringRedisRepository.hasKey(key)) {
+			stringRedisRepository.delete(key);
+		}
+
+		if (stringRedisRepository.hasKey(stockKey)) {
+			stringRedisRepository.delete(stockKey);
+		}
 	}
 
 	@DisplayName("레디스에 문자열 데이터가 성공적으로 저장된다. - Void")
 	@Test
-	void stringRedisRepository_save() {
+	void save_success() {
 		// Then
 		assertThat(stringRedisRepository.get(key)).isEqualTo(value);
 	}
 
-	@DisplayName("레디스의 특정 데이터가 성공적으로 삭제된다. - Void")
-	@Test
-	void stringRedisRepository_delete() {
-		// When
-		stringRedisRepository.delete(key);
-
-		// Then
-		assertThat(stringRedisRepository.hasKey(key)).isFalse();
-	}
-
 	@DisplayName("레디스의 특정 데이터가 성공적으로 조회된다. - String(Value)")
 	@Test
-	void stringRedisRepository_get() {
+	void get_success() {
 		// When
 		String actual = stringRedisRepository.get(key);
 
@@ -62,8 +59,28 @@ class StringRedisRepositoryTest {
 
 	@DisplayName("레디스의 특정 데이터 존재 여부를 성공적으로 체크한다. - Boolean")
 	@Test
-	void stringRedisRepository_hasKey() {
+	void hasKey_success() {
 		// When & Then
 		assertThat(stringRedisRepository.hasKey("not found key")).isFalse();
+	}
+
+	@DisplayName("레디스의 특정 데이터가 성공적으로 삭제된다. - Void")
+	@Test
+	void delete_success() {
+		// When
+		stringRedisRepository.delete(key);
+
+		// Then
+		assertThat(stringRedisRepository.hasKey(key)).isFalse();
+	}
+
+	@DisplayName("레디스의 특정 데이터의 값이 1 증가한다.")
+	@Test
+	void increment_success() {
+		// When
+		Long actual = stringRedisRepository.increment(stockKey);
+
+		// Then
+		assertThat(actual).isEqualTo(1L);
 	}
 }
